GPU SpTRSV fp16 storage: default ON on the fp32 build's kernel backends; fix the fp32 inv_scale^2 overflow behind the silent-NaN solves
DEFAULT FLIP (was opt-in): APXCHOL_GPU_SPTRSV_FP16 unset now resolves ON
whenever a kernel backend (dataflow / level-set) is the resolved choice on
the fp32 SpTRSV build -- measured -12% gpu_pcg_loop on iter0040, -4..8%
grids, within +-7% elsewhere, -15..20% device factor bytes, iteration-
neutral on all 9 suite matrices. Explicit =0 opts out; the fp64 build's
AUTO (cuSPARSE-if-fits / level-set) and a forced =cusparse keep fp16 OFF.
fp16_env_tristate() / fp16_resolved() carry the resolution; the solve.cpp
banner reads the same function.

The flip exposed a PRE-EXISTING fp16 defect (verified: HEAD~ with explicit
=1 fails identically): setup pre-squared the back solve's per-column input
scale into fp32 -- inv_scale2[j] = fp32((1/s_j)^2) -- which overflows to
+Inf for column scales s_j < ~5.4e-20, and the Inf dissolved the whole PCG
solve into NaN with no error anywhere. Such a column does occur on real
factors: PcgFusion.RepeatedSolvesBitIdenticalSDDM's sddm_grid(100,100,0.5)
factor at OMP_NUM_THREADS=4/6/8 has one with s_j ~ 1e-29 (the thread count
perturbs the sampled factor by merge-order ulps; T=1/2/3/5/12/16 do not
produce one). The "fails under ctest -j16, passes alone" symptom was a
confound: ctest pins OMP_NUM_THREADS=4 per test, the raw binary ran at 16
-- the failure is deterministic solo at T=4 and has nothing to do with GPU
sharing (0 failures in 20 runs under 8-15 concurrent GPU processes at
other thread counts; both kernel backends failed identically).

Fix, mirroring the CPU FP16_SCALED contract (omp.h back_dir::rhs already
forms x_in * r^2 in double): d_inv_scale2_ is now DOUBLE on the device
(r_j^2 exact) and both fp16 kernels form rhs * r^2 in double with one
narrowing cast -- the product is small whenever the true solution is
representable; only the pre-squared scale was not. Two guards on top:
a column whose scale cannot be represented at all (fp32 1/s_j or
fp32(L_jj)/s_j overflows) falls back to s_j = 1, its off-diagonals --
negligible next to that diagonal at any fp32 precision -- drop/flush away;
and setup() verifies every diag / inv_scale^2 is finite (diag nonzero) and
THROWS otherwise: a factor the fp16 storage cannot represent fails loudly,
never as NaN.

Tests updated for the runtime default:
  * GpuDataflow.PairThroughCudaSptrsvMatchesLevelSetAndIsDeterministic pins
    APXCHOL_GPU_SPTRSV_FP16 both ways ("=0" explicitly for the fp32
    sub-case; unset now resolves fp16 and compared fp16 output against the
    fp32 expectations).
  * CustomEliminator.ExactCliqueFactorIsExact runs both ways on the CUDA
    build: "exact factor => <= 3 iterations" requires lossless storage;
    the fp16 sub-case allows the one extra iteration its 2^-11 rounding
    costs.

Validated: full ctest -j16 x5 = 189/189 each; the three affected tests x30
solo (ambient and OMP_NUM_THREADS=4) = 0 failures; CPU build 187/187;
compute-sanitizer memcheck clean on the fp16 pair; fp64 scratch build
runtime-checked (unset -> fp64 level-set AUTO unchanged; =1 -> fp16
level-set converges).

+++ include/apxchol/solver/sptrsv/cuda_dataflow.h
@@ -131,10 +131,11 @@ void dataflow_solve(cudaStream_t stream, int m, bool reverse,
 /// The same sweep on the fp16 per-column-scaled storage (cuda_host.h
 /// contract; identical semantics to levelset_solve_fp16: `vals16` binary16
 /// bit patterns, the diagonal slot skipped, divide by the fp32 `diag[i]`,
-/// rhs[i] scaled once by `in_scale[i]` when non-null).
+/// rhs[i] scaled once by the DOUBLE `in_scale[i]` when non-null -- r_i^2
+/// exact, product in double, one narrowing cast; see cuda_levelset.h).
 void dataflow_solve_fp16(cudaStream_t stream, int m, bool reverse,
                          const int* rowptr, const int* colidx, const unsigned short* vals16,
-                         const float* diag, const float* in_scale,
+                         const float* diag, const double* in_scale,
                          const int* batch_start, int n_batches,
                          unsigned long long* tag, unsigned epoch, int* ctrl, int grid,
                          const sptrsv_gpu_value_t* rhs, sptrsv_gpu_value_t* out);
